refactor: create validate_routes.py
